Require an exact match for session names when attaching
...otherwise tmux will allow prefix matches, e.g. an existing session
'foobar' will be attached when session 'foo' is requested.

In the context of sshmux, prefix matching doesn't make sense. If the
session doesn't exist sshmux goes on to create it with the given session
name, so users won't be able to reliably use the prefix matching
functionality anyway.

The quotation marks are to avoid zsh '=' expansion; see
http://zsh.sourceforge.net/Doc/Release/Expansion.html#g_t_0060_003d_0027-expansion